+++ internal/check/uuid.go
@@ -6,6 +6,7 @@ import (
 	"github.com/google/uuid"
 )
 
+// ValidateUUID returns error on wrong UUID
 func ValidateUUID(id string) error {
 
 	_, err := uuid.Parse(strings.TrimSpace(id))