+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [v1.0.7] - 2026-09-
+### Added
+- Parse: add prefix before tag
+- Conf: check TLS certificate of a host through proxy
+
 ## [v1.0.6] - 2026-09-10
 ### Added
 - Validate `UUID` for `vless` and `tuic`